Update issue 256
アクティビティスキーマの調整を行いました。


git-svn-id: http://aipo.googlecode.com/svn/trunk@746 b7d4e767-949e-c8c6-5735-6c11b8a2f74f

+++ orm/src/main/java/com/aipo/orm/model/social/auto/_ActivityMap.java
@@ -37,11 +37,11 @@ public abstract class _ActivityMap extends CayenneDataObject {
 
     public static final String ID_PK_COLUMN = "ID";
 
-    public void setIsRead(Short isRead) {
+    public void setIsRead(Integer isRead) {
         writeProperty("isRead", isRead);
     }
-    public Short getIsRead() {
-        return (Short)readProperty("isRead");
+    public Integer getIsRead() {
+        return (Integer)readProperty("isRead");
     }
 
     public void setLoginName(String loginName) {
